+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Add ability to refresh project after changes in dependencies
+
 ## [0.1.8] - 2023-04-12
 
 * Fix invalid boundary resolving of composite steps in composite files